requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
Flask-Script
Flask-WTF
Jinja2
Mako
MarkupSafe
SQLAlchemy
WTForms
Werkzeug
alembic
argparse
bbcode
ipython
itsdangerous
passlib
psycopg2
python-dateutil==1.5
wsgiref
